Refactor : newJourney.js -> NewJourneyContext.jsx 이름 변경
