+++ client/project-there/src/components/journeyUpload/PeopleCounter.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import {
   useNewJourneyActions,
   useNewJourneyValue,
-} from '../../contexts/newJourney';
+} from '../../contexts/NewJourneyContext';
 
 const Wrapper = styled.div`
   margin-top: 16px;
