+++ client/project-there/src/components/journeyUpload/SchedulePicker.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import {
   useNewJourneyActions,
   useNewJourneyValue,
-} from '../../contexts/newJourney';
+} from '../../contexts/NewJourneyContext';
 import DatePickerContainer from './DatePickerContainer';
 import ScheduleModal from './ScheduleModal';
