+++ client/project-there/src/components/journeyUpload/ThemeButton.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { useNewJourneyValue } from '../../contexts/newJourney';
+import { useNewJourneyValue } from '../../contexts/NewJourneyContext';
 
 const Button = styled.button`
   &:hover {
