+++ client/project-there/src/pages/JourneyUpdatePage.jsx
@@ -1,21 +1,15 @@
 import React, { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import JourneyUpdateContainer from '../components/journeyUpdate/JourneyUpdateContainer';
-import JourneyUploadContainer from '../components/journeyUpload/JourneyUploadContainer';
 import { useAuthActions } from '../contexts/auth';
-import JourneyDetailProvider, {
-  useJourneyDetailValue,
-} from '../contexts/journeyDetail';
-import NewJourneyProvider, {
-  useNewJourneyActions,
-} from '../contexts/newJourney';
+import NewJourneyProvider from '../contexts/NewJourneyContext';
 import { getAccessToken } from '../lib/utills/controlAccessToken';
 
 function JourneyUpdatePage() {
   const params = useParams();
 
   // 유저 인증
-  const { authenticateUser, initUserInfo } = useAuthActions();
+  const { authenticateUser } = useAuthActions();
 
   useEffect(() => {
     // 마운트 될 때 accessToken으로 받은 유저 정보 업데이트
@@ -27,11 +21,9 @@ function JourneyUpdatePage() {
   }, []);
 
   return (
-    <JourneyDetailProvider>
-      <NewJourneyProvider>
-        <JourneyUpdateContainer journeyId={params.journeyId} />
-      </NewJourneyProvider>
-    </JourneyDetailProvider>
+    <NewJourneyProvider>
+      <JourneyUpdateContainer journeyId={params.journeyId} />
+    </NewJourneyProvider>
   );
 }
